Add files via upload
This was a university project that I did in my final semester of University in 2019 while I was studying at University of Murdoch, Singapre campus. 
Me and my group build this software. I was the project lead and the software architecture of this development.
Meanwhile, I also contributed extensively to the code by writing the backend logic of the SQL database, the interaction between the PHP and SQL database, the API page between Google Map and our website.

The requirements of this project as here:

This project seeks to develop interpretation material, including traditional signage and a mobile phone app, for the Murdoch University Banksia Court. 

This specialised banksia arboretum was established in 1995 to mark the 20th anniversary of the opening of the University. It contains 25 species of banksias and provides a rich food source for honeyeaters and black cockatoos, while the dense undergrowth provides vital habitat for ‘Conservation Dependent’ Southern Brown Bandicoots (Quendas), native lizards and invertebrates. Murdoch’s Banksia Court is the only other dedicated Banksia Garden in WA apart from the one in King’s Park. 

The project’s aims are: 
- Develop interactive technological solutions to signage that educate and inspire users about banksias and their environmental and cultural importance. 
- Create interpretive signage for the site, including European and Nyungar information (name, flowering time, place of origin, etc.). 


How to setup the website -- User friendly guide
1. Download Installation Manual
2. Download Rar File from Google-link to setup the web development - https://drive.google.com/file/d/1t-OpEXKRIxWDOwUqZiEe6zIkN025vn2S/view?usp=drive_link